test(test-optimization): centralize browser RUM boundary tests
Playwright and Selenium RUM boundary coverage used standalone instrumentation jobs while Cypress lived with its framework integration suite. Keep each framework's behavior and failure cases under its existing integration runner.
